allow skipping package test; improve cleanup

+++ .scripts/test_package.sh
@@ -40,64 +40,63 @@ else
 fi
 TYPE="_${OS}_$(uname -m)"
 PACKAGE_NAME=fstar_$CURRENT_VERSION$TYPE
-git clean -ffdx src/ocaml-output release
-PACKAGE_NAME=$PACKAGE_NAME make -j6 package
-cd src/ocaml-output
-
-diag "*** Unzip and verify the Package  ***"
-mkdir "$FSTAR_HOST_HOME/release"
-rm -rf /tmp/fstar_package
-mkdir /tmp/fstar_package
-cd /tmp/fstar_package
 BUILD_PACKAGE_FILENAME=$PACKAGE_NAME$ext
 BUILD_PACKAGE="$FSTAR_HOST_HOME/src/ocaml-output/$BUILD_PACKAGE_FILENAME"
-if [[ -f $BUILD_PACKAGE ]] ; then
-  if [[ $ext = .zip ]] ; then
-      unzip -o $BUILD_PACKAGE
-  else
-      tar xf $BUILD_PACKAGE
-  fi
-else
+rm -rf "$BUILD_PACKAGE" src/ocaml-output/fstar
+PACKAGE_NAME=$PACKAGE_NAME make -j6 package
+
+diag "*** Check that the package was created  ***"
+if ! [[ -f $BUILD_PACKAGE ]] ; then
   echo -e "src/ocaml-output/make package did not create ${BUILD_PACKAGE_FILENAME}"
   echo -e "* ${RED}FAIL!${NC}"
   exit 1
 fi
-NEW_BUILD_PACKAGE="$FSTAR_HOST_HOME/release/$BUILD_PACKAGE_FILENAME"
-cp "$BUILD_PACKAGE" "$NEW_BUILD_PACKAGE"
-BUILD_PACKAGE="$NEW_BUILD_PACKAGE"
 
-diag "*** Test the binary package ***"
-
-# We need two FSTAR_HOMEs in this script: one for the host (from where
-# we build F*) and one for the package (from where we test the
-# obtained binary). FSTAR_HOME is the latter.
+diag "*** Unzip the binary package ***"
+rm -rf /tmp/fstar_package
+mkdir /tmp/fstar_package
+cd /tmp/fstar_package
+if [[ $ext = .zip ]] ; then
+    unzip -o $BUILD_PACKAGE
+else
+    tar xf $BUILD_PACKAGE
+fi
 cd fstar
-export FSTAR_HOME="$PWD"
-
-# Move doc and examples to the tmp directory.
-# We move them elsewhere since we
-# don't want to rely on relative paths in their Makefiles.
-rm -rf /tmp/fstar_examples /tmp/fstar_doc
-mv share/fstar/examples /tmp/fstar_examples
-mv share/fstar/doc /tmp/fstar_doc
 
 diag "-- Versions --"
 bin/fstar.exe --version
 bin/z3 --version
 
-diag "-- Verify all examples --"
-make -j6 -C /tmp/fstar_examples && make -j6 -C /tmp/fstar_doc/tutorial regressions
-if [ $? -ne 0 ]; then
-  echo -e "* ${RED}FAIL!${NC} for all examples - make returned $?"
-  exit 1
-else
-  echo -e "* ${GREEN}PASSED!${NC} for all examples"
+if [[ -z "$FSTAR_SKIP_PACKAGE_TEST" ]] ; then
+
+    diag "*** Test the binary package"
+    # We need two FSTAR_HOMEs in this script: one for the host (from where
+    # we build F*) and one for the package (from where we test the
+    # obtained binary). FSTAR_HOME is the latter.
+    export FSTAR_HOME="$PWD"
+
+    # Move doc and examples to the tmp directory.
+    # We move them elsewhere since we
+    # don't want to rely on relative paths in their Makefiles.
+    rm -rf /tmp/fstar_examples /tmp/fstar_doc
+    mv share/fstar/examples /tmp/fstar_examples
+    mv share/fstar/doc /tmp/fstar_doc
+
+    diag "-- Verify all examples --"
+    make -j6 -C /tmp/fstar_examples && make -j6 -C /tmp/fstar_doc/tutorial regressions
+    if [ $? -ne 0 ]; then
+        echo -e "* ${RED}FAIL!${NC} for all examples - make returned $?"
+        exit 1
+    else
+        echo -e "* ${GREEN}PASSED!${NC} for all examples"
+    fi
+
 fi
 
 cd $FSTAR_HOST_HOME
 
 # Cleanup
-rm -rf /tmp/fstar_examples /tmp/fstar_doc
+rm -rf /tmp/fstar_examples /tmp/fstar_doc src/ocaml-output/fstar
 
 # From this point on, we should no longer need FSTAR_HOME.
 export FSTAR_HOME=
